First Draft ready

+++ frontend/src/components/Calculator/Calculate.jsx
@@ -29,7 +29,7 @@ const Calculate = () => {
     // emittedCO2 = emittedCO2 / 1000;
     console.log(emitted);
     const date = values.date;
-    const emittedCO2 = parseInt(emitted);
+    const emittedCO2 = parseFloat(emitted);
     console.log(emittedCO2);
     const body = { emittedCO2, date };
     const response = await manageCo2("/auth/addCo2", body, token);
